requirements.txt
--extra-index-url https://download.pytorch.org/whl/nightly/cu121           
# Had to use better version as older one was not compatible with RTX PRO 4500 Blackwell
torch
torchvision

h5py==3.1.0
imageio==2.9.0
Markdown==3.3.4
mat73==0.58
matplotlib==3.3.4
MedPy==0.5.2
numpy==1.26.4
opencv-python==4.5.3.56
panda==0.3.1
pandas==1.1.5
Pillow==9.5.0
pyparsing==2.4.7
scikit-image==0.19.3
scikit-learn==1.2.2
scipy==1.10.1
timm==0.4.12
tqdm==4.62.2

# conda install -c conda-forge pydensecrf

# pip uninstall numpy scipy scikit-learn -y
# pip install numpy==1.26.4 scipy==1.10.1 scikit-learn==1.2.2